Dodavanje nove Sobe- Http POST

+++ BACKEND/Models/Soba.cs
@@ -13,6 +13,6 @@ namespace BACKEND.Models
         public string? Dostupnost { get; set; }
         
         [Column(name: "broj_sobe")]
-        public double? BrojSobe { get; set; }
+        public int? BrojSobe { get; set; }
     }
 }
